requirements.txt
django==4.2
djangorestframework
django-cors-headers
tensorflow-cpu==2.16.1
pillow
numpy==1.23.5
gunicorn